rename kernel_manager to spice_utils
